# src/apidocs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "O-Horizon Recommendation APIs",
    "version": "1.0.0",
    "contact": {
      "name": "Joffin Joy",
      "url": "https://www.linkedin.com/in/joffin-joy/"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3020"
    }
  ],
  "paths": {
...
    "/recommendation/get-collaborative-recommendations": {
      "post": {
        "tags": [
          "Recommendation"
        ],
        "summary": "Get collaborative Recommendations",
        "description": "Get item recommendations for a user based on their similarities with other users (Collaborative Filtering)",
        "operationId": "getCollaborativeRecommendations",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "userId": {
                    "type": "string",
                    "example": "645a30763153bfd734ceebb0"
                  }
                }
              },
              "example": {
                "userId": "645a30763153bfd734ceebb0"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Recommendations successfully retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Recommendations For User: 645a30763153bfd734ceebb0 Retrieved"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "itemId": {
                            "type": "string",
                            "example": "645a868e0b03a39d412c8ada"
                          },
                          "title": {
                            "type": "string",
                            "example": "Enhancing Student Learning through Peer Feedback ClusterNumber7"
                          }
                        }
                      }
                    }
                  },
                  "example": {
                    "status": true,
                    "message": "Recommendations For User: 645a30763153bfd734ceebb0 Retrieved",
                    "data": [
                      {
                        "itemId": "645a868e0b03a39d412c8ada",
                        "title": "Enhancing Student Learning through Peer Feedback ClusterNumber7"
                      },
                      {
                        "itemId": "645a868eba244d094e865993",
                        "title": "Effective Approaches to Differentiated Learning ClusterNumber7"
                      },
                      {
                        "itemId": "645a868eef45797fca900cc0",
                        "title": "Effective Differentiation for Students with Autism Spectrum Disorders ClusterNumber7"
                      },
                      {
                        "itemId": "645a868ea39a294b2f548575",
                        "title": "Strategies for Supporting Students with Anxiety in the Classroom ClusterNumber7"
                      },
                      {
                        "itemId": "645a868e4b9dfb73b63ca44c",
                        "title": "Supporting Students with Autism in the Classroom ClusterNumber7"
                      }
                    ]
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/recommendation/get-item-page-recommendations": {
      "post": {
        "tags": [
          "Recommendation"
        ],
        "summary": "Get Item Page Recommendations",
        "description": "Get a list of recommended items which are similar to a selected item",
        "operationId": "getItemPageRecommendations",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "itemId": {
                    "type": "string",
                    "example": "645a868ea39a294b2f548575"
                  }
                }
              },
              "example": {
                "itemId": "645a868ea39a294b2f548575"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Item Page Recommendations Fetched",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "boolean",
                      "description": "Status of the response.",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "description": "A message describing the status of the response.",
                      "example": "Item Page Recommendations Fetched"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "itemId": {
                            "type": "string",
                            "description": "ID of the item.",
                            "example": "645a868e55ae998e9e01288a"
                          },
                          "title": {
                            "type": "string",
                            "description": "Title of the item.",
                            "example": "Strategies for Supporting Students with Behavioral Challenges ClusterNumber3"
                          }
                        }
                      },
                      "example": [
                        {
                          "itemId": "645a868e55ae998e9e01288a",
                          "title": "Strategies for Supporting Students with Behavioral Challenges ClusterNumber3"
                        },
                        {
                          "itemId": "645a868e4b9dfb73b63ca44c",
                          "title": "Supporting Students with Autism in the Classroom ClusterNumber7"
                        },
                        {
                          "itemId": "645a868ee10b8ea4ba7ee00c",
                          "title": "Effective Classroom Management for Students with Anxiety ClusterNumber6"
                        },
                        {
                          "itemId": "645a868e6e7ec1d4955bb4b5",
                          "title": "Supporting Students with Depression in the Classroom ClusterNumber1"
                        },
                        {
                          "itemId": "645a868ee6d0d28d4ad3104e",
                          "title": "Supporting Students with Dyscalculia in the Classroom ClusterNumber9"
                        },
                        {
                          "itemId": "645a868eb607cba4799038ef",
                          "title": "Supporting Students with ADHD in the Virtual Classroom ClusterNumber9"
                        }
                      ]
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/recommendation/get-profile-page-recommendations": {
      "post": {
        "tags": [
          "Recommendation"
        ],
        "summary": "Get Profile Page Recommendations",
        "description": "Get a list of items which are similar to items which the user rated previously",
        "operationId": "getProfilePageRecommendations",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "userId": {
                    "type": "string",
                    "example": "645a868e05ed75e9975a0aa3"
                  }
                }
              },
              "example": {
                "userId": "645a868e05ed75e9975a0aa3"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Profile page recommendations fetched successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "boolean",
                      "description": "Indicates whether the request was successful or not."
                    },
                    "message": {
                      "type": "string",
                      "description": "A message describing the outcome of the request."
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "itemId": {
                            "type": "string",
                            "description": "The unique identifier of the recommended item."
                          },
                          "title": {
                            "type": "string",
                            "description": "The title of the recommended item."
                          },
                          "score": {
                            "type": "number",
                            "format": "float",
                            "description": "The score indicating how well the recommended item matches the user's preferences."
                          }
                        }
                      }
                    }
                  }
                },
                "example": {
                  "status": true,
                  "message": "Profile Page Recommendations Fetched",
                  "data": [
                    {
                      "itemId": "645a868eef40453a73c3b2f9",
                      "title": "Effective Strategies for Classroom Discipline ClusterNumber2",
                      "score": 1.625
                    },
                    {
                      "itemId": "645a868e8dce646d34611481",
                      "title": "Effective Classroom Management for Remote Learning ClusterNumber10",
                      "score": 1.525
                    },
                    {
                      "itemId": "645a868e09d47ec3875d70a1",
                      "title": "Making the Most of Classroom Technology ClusterNumber2",
                      "score": 1.525
                    },
                    {
                      "itemId": "645a868e10b5934aa265c41a",
                      "title": "Making the Most of Classroom Libraries ClusterNumber6",
                      "score": 1.525
                    },
                    {
                      "itemId": "645a868e31d8251630e58a9a",
                      "title": "Collaborative Learning in the Classroom ClusterNumber6",
                      "score": 1.525
                    },
                    {
                      "itemId": "645a868e542b0e2567d12fc4",
                      "title": "Promoting Social Emotional Learning in the Classroom ClusterNumber2",
                      "score": 1.525
                    },
                    {
                      "itemId": "645a868ed2b024f3fbf284b8",
                      "title": "Differentiated Instruction for Students with Chronic Illnesses ClusterNumber8",
                      "score": 1.25
                    },
                    {
                      "itemId": "645a868eae132211d46ba383",
                      "title": "Differentiated Instruction for Students with Giftedness ClusterNumber7",
                      "score": 1.25
                    },
                    {
                      "itemId": "645a868e42be65b605a66962",
                      "title": "Differentiated Instruction for Students with Emotional Disturbance ClusterNumber2",
                      "score": 1.25
                    },
                    {
                      "itemId": "645a868e8fd03d56f2f579ac",
                      "title": "Differentiated Instruction for Students with Trauma ClusterNumber1",
                      "score": 1.25
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}